restructured workout navigation
